Function async_std_resolver::resolver

pub async fn resolver(
    config: ResolverConfig, 
    options: ResolverOpts
) -> Result<AsyncStdResolver, ResolveError>

Construct a new async-std based AsyncResolver with the provided configuration.

Arguments

config - configuration, name_servers, etc. for the Resolver

options - basic lookup options for the resolver

Returns

A tuple containing the new AsyncResolver and a future that drives the background task that runs resolutions for the AsyncResolver. See the documentation for AsyncResolver for more information on how to use the background future.